requirements.txt
scipion-pyworkflow